The update.py file has been split into modules
